requirements.txt
aioredis==1.3.1
amqp==2.5.2
appdirs==1.4.3
asgiref==3.2.3
aspy.yaml==1.3.0
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.1.3
Automat==0.8.0
billiard==3.6.2.0
black==19.10b0
celery==4.4.0
certifi==2019.11.28
cffi==1.13.2
cfgv==2.0.1
channels==2.4.0
channels-redis==2.4.1
Click==7.0
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
Django==3.0.2
django-cors-headers==3.2.1
django-environ==0.4.5
django-filter==2.2.0
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
hiredis==1.0.1
hyperlink==19.0.0
identify==1.4.11
idna==2.8
importlib-metadata==1.5.0
incremental==17.5.0
kombu==4.6.7
Markdown==3.1.1
mccabe==0.6.1
msgpack==0.6.2
nodeenv==1.3.4
pathspec==0.7.0
pre-commit==2.0.1
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyHamcrest==2.0.0
pyOpenSSL==19.1.0
pytz==2019.3
PyYAML==5.3
regex==2020.1.8
sentry-sdk==0.14.1
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
toml==0.10.0
Twisted==19.10.0
txaio==20.1.1
typed-ast==1.4.1
urllib3==1.25.8
vine==1.3.0
virtualenv==16.7.9
zipp==2.1.0
zope.interface==4.7.1